Moved logging package into transport package. #233
