+++ reactor-netty-core/src/main/java/reactor/netty/transport/logging/ReactorNettyLoggingHandler.java
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package reactor.netty.logging;
+package reactor.netty.transport.logging;
 
 import java.nio.charset.Charset;
