+++ reactor-netty-core/src/test/java/reactor/netty/transport/logging/AdvancedByteBufFormatTest.java
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package reactor.netty.logging;
+package reactor.netty.transport.logging;
 
 import static org.assertj.core.api.Assertions.assertThat;
